Crate mistralrs_core

Modules

layers

Structs

AnyMoeConfig

AnyMoeLoader

AnyMoePipeline

ChatCompletionChunkResponse

Chat completion streaming request chunk.

ChatCompletionResponse

An OpenAI compatible chat completion response.

ChatTemplate

Template for chat models including bos/eos/unk as well as the chat template.

Choice

Chat completion choice.

ChunkChoice

Chat completion streaming chunk choice.

CompletionChoice

Completion request choice.

CompletionChunkChoice

Chat completion streaming chunk choice.

CompletionChunkResponse

Completion request choice.

CompletionResponse

An OpenAI compatible completion response.

Delta

Delta in content for streaming response.

DeviceLayerMapMetadata

DeviceMapMetadata

Metadata to initialize the device mapper.

GGMLLoader

A loader for a GGML model.

GGMLLoaderBuilder

A builder for a GGML loader.

GGMLSpecificConfig

Config for a GGML loader.

GGUFLoader

Loader for a GGUF model.

GGUFLoaderBuilder

A builder for a GGUF loader.

GGUFSpecificConfig

A config for a GGUF loader.

GemmaLoader

NormalLoader for a Gemma model.

Idefics2Loader

VisionLoader for an Idefics 2 Vision model.

LLaVALoader

[VisionLoader] for an LLaVA Vision model.

LLaVANextLoader

[VisionLoader] for an LLaVANext Vision model.

LayerDeviceMapper

A device mapper which does device mapping per hidden layer.

LlamaLoader

NormalLoader for a Llama model.

LoaderBuilder

A builder for a loader using the selected model.

LocalModelPaths

All local paths and metadata necessary to load a model.

Logprobs

Logprobs per token.

MemoryUsage

MistralLoader

MistralRs

The MistralRs struct handles sending requests to the engine. It is the core multi-threaded component of mistral.rs, and uses mspc Sender and Receiver primitives to send and receive requests to the engine.

MistralRsBuilder

The MistralRsBuilder takes the pipeline and a scheduler method and constructs an Engine and a MistralRs instance. The Engine runs on a separate thread, and the MistralRs instance stays on the calling thread.

MixtralLoader

NormalLoader

A loader for a “normal” (non-quantized) model.

NormalLoaderBuilder

A builder for a loader for a “normal” (non-quantized) model.

NormalRequest

A normal request request to the MistralRs

NormalSpecificConfig

Config specific to loading a normal model.

Ordering

Adapter model ordering information.

Phi2Loader

NormalLoader for a Phi 2 model.

Phi3Loader

NormalLoader for a Phi 3 model.

Phi3VLoader

VisionLoader for a Phi 3 Vision model.

Qwen2Loader

NormalLoader for a Qwen 2 model.

ResponseLogprob

A logprob with the top logprobs for this token.

ResponseMessage

Chat completion response message.

SamplingParams

Sampling params are used to control sampling.

SpeculativeConfig

Metadata for a speculative pipeline

SpeculativeLoader

A loader for a speculative pipeline using 2 Loaders.

SpeculativePipeline

Speculative decoding pipeline: https://arxiv.org/pdf/2211.17192

Starcoder2Loader

NormalLoader for a Starcoder2 model.

TopLogprob

Top-n logprobs element

Usage

OpenAI compatible (superset) usage during a request.

VisionLoader

A loader for a vision (non-quantized) model.

VisionLoaderBuilder

A builder for a loader for a vision (non-quantized) model.

VisionSpecificConfig

Config specific to loading a vision model.

Enums

AnyMoeExpertType

Constraint

Control the constraint with Regex or Yacc.

GGUFArchitecture

MistralRsError

ModelDType

DType for the model.

ModelKind

The kind of model to build.

ModelSelected

NormalLoaderType

The architecture to load the normal model as.

Request

A request to the Engine, encapsulating the various parameters as well as the mspc response Sender used to return the Response.

RequestMessage

Message or messages for a Request.

Response

The response enum contains 3 types of variants:

SchedulerMethod

The scheduler method controld how sequences are scheduled during each step of the engine. For each scheduling step, the scheduler method is used if there are not only running, only waiting sequences, or none. If is it used, then it is used to allow waiting sequences to run.

StopTokens

Stop sequences or ids.

TokenSource

The source of the HF token.

VisionLoaderType

The architecture to load the vision model as.

Constants

SYSTEM_FINGERPRINT

Statics

TERMINATE_ALL_NEXT_STEP

Terminate all sequences on the next scheduling step. Be sure to reset this.

Traits

Loader

The Loader trait abstracts the loading process. The primary entrypoint is the load_model method.

ModelPaths

ModelPaths abstracts the mechanism to get all necessary files for running a model. For example LocalModelPaths implements ModelPaths when all files are in the local file system.

Pipeline

TryIntoDType

Type which can be converted to a DType

VisionModelLoader

Functions

get_model_dtype

get_tgt_non_granular_index

get_toml_selected_model_dtype

initialize_logging

This should be called to initialize the debug flag and logging. This should not be called in mistralrs-core code due to Rust usage.

Type Aliases

MessageContent