Enum mistralrs_core::ModelSelected

pub enum ModelSelected {
    Plain {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
        arch: NormalLoaderType,
    },
    XLora {
        model_id: Option<String>,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        repeat_last_n: usize,
        order: String,
        tgt_non_granular_index: Option<usize>,
        arch: NormalLoaderType,
    },
    Lora {
        model_id: Option<String>,
        tokenizer_json: Option<String>,
        adapters_model_id: String,
        repeat_last_n: usize,
        order: String,
        arch: NormalLoaderType,
    },
    GGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
    },
    XLoraGGUF {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    LoraGGUF {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
        adapters_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    GGML {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
        gqa: usize,
    },
    XLoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        gqa: usize,
    },
    LoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        repeat_last_n: usize,
        adapters_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        gqa: usize,
    },
}

Variants

Plain

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

arch: NormalLoaderType

The architecture of the model.

Select a plain model, without quantization or adapters

XLora

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

arch: NormalLoaderType

The architecture of the model.

Select an X-LoRA architecture

Lora

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

adapters_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

arch: NormalLoaderType

The architecture of the model.

Select a LoRA architecture

GGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select a GGUF model.

XLoraGGUF

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

Select a GGUF model with X-LoRA.

LoraGGUF

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

adapters_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

Select a GGUF model with LoRA.

GGML

Fields

tok_model_id: String

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

gqa: usize

GQA value

Select a GGML model.

XLoraGGML

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

gqa: usize

GQA value

Select a GGML model with X-LoRA.

LoraGGML

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename, only applicable if quantized is set. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

adapters_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

gqa: usize

GQA value

Select a GGML model with LoRA.