mistralrs_core

Trait VisionPromptPrefixer

pub trait VisionPromptPrefixer: Send + Sync {
    // Required method
    fn prefix_image(&self, image_index: usize, prompt: &str) -> String;
}

Prepend a vision tag appropriate for the model to the prompt. Image indexing is assumed that start at

Required Methods

fn prefix_image(&self, image_index: usize, prompt: &str) -> String

Prefix for inclusion in messages (may do nothing if the chat template handles it).