mistralrs_core

Type Alias SearchCallback

pub type SearchCallback = dyn Fn(&SearchFunctionParameters) -> Result<Vec<SearchResult>> + Send + Sync;

Callback used to override how search results are gathered. The returned vector must be sorted in decreasing order of relevance.