Crate mistralrs_core

Macros

deserialize_chat_template

get_bias_if_not_allowed

get_mut_arcmutex

get_mut_group

get_paths

handle_pipeline_forward_error

handle_seq_error

handle_seq_error_ok

handle_seq_error_stateaware

handle_seq_error_stateaware_ok

lora_model_loader

normal_model_loader

xlora_model_loader

Structs

ChatCompletionChunkResponse

ChatCompletionResponse

An OpenAI compatible chat completion response.

Choice

ChunkChoice

CompletionChoice

CompletionResponse

An OpenAI compatible completion response.

Delta

DeviceMapMetadata

GGMLLoader

GGMLLoaderBuilder

A builder for a GGML loader.

GGMLSpecificConfig

Config for a GGML loader.

GGUFLoader

GGUFLoaderBuilder

A builder for a GGUF loader.

GGUFSpecificConfig

A config for a GGUF loader.

GemmaLoader

LayerDeviceMapper

LlamaLoader

LoaderBuilder

Logprobs

MistralLoader

MistralRs

The MistralRs struct handles sending requests to the engine. It is the core multi-threaded component of mistral.rs, and uses mspc Sender and Receiver primitives to send and receive requests to the engine.

MistralRsBuilder

The MistralRsBuilder takes the pipeline and a scheduler method and constructs an Engine and a MistralRs instance. The Engine runs on a separate thread, and the MistralRs instance stays on the calling thread.

MixtralLoader

NormalLoader

A loader for a “normal” (non-quantized) model.

NormalLoaderBuilder

A builder for a loader for a “normal” (non-quantized) model.

NormalSpecificConfig

Config specific to loading a normal model.

Ordering

Phi2Loader

Request

A request to the Engine, encapsulating the various parameters as well as the mspc response Sender used to return the Response.

ResponseLogprob

ResponseMessage

SamplingParams

Sampling params are used to control sampling.

TopLogprob

Usage

OpenAI compatible (superset) usage during a request.

Enums

Constraint

Control the constraint with Regex or Yacc.

ModelKind

The kind of model to build.

ModelSelected

NormalLoaderType

The architecture to load the normal model as.

RequestMessage

Message or messages for a Request.

Response

The response enum contains 3 types of variants:

SchedulerMethod

The scheduler method controld how sequences are scheduled during each step of the engine. For each scheduling step, the scheduler method is used if there are not only running, only waiting sequences, or none. If is it used, then it is used to allow waiting sequences to run.

StopTokens

Stop sequences or ids.

TokenSource

The source of the HF token.

Constants

SYSTEM_FINGERPRINT

Traits

Loader

The Loader trait abstracts the loading process. The primary entrypoint is the load_model method.

Pipeline

Functions

get_tgt_non_granular_index