mistralrs_core

Enum ModelSelected

pub enum ModelSelected {

    Toml {
        file: String,
    },
    Run {

        model_id: String,
        tokenizer_json: Option<String>,
        dtype: ModelDType,
        topology: Option<String>,
        organization: Option<IsqOrganization>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        imatrix: Option<PathBuf>,
        calibration_file: Option<PathBuf>,
        max_edge: Option<u32>,
        max_seq_len: usize,
        max_batch_size: usize,
        max_num_images: Option<usize>,
        max_image_length: Option<usize>,
        hf_cache_path: Option<PathBuf>,
        matformer_config_path: Option<PathBuf>,
        matformer_slice_name: Option<String>,
    
},
    Plain {

        model_id: String,
        tokenizer_json: Option<String>,
        arch: Option<NormalLoaderType>,
        dtype: ModelDType,
        topology: Option<String>,
        organization: Option<IsqOrganization>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        imatrix: Option<PathBuf>,
        calibration_file: Option<PathBuf>,
        max_seq_len: usize,
        max_batch_size: usize,
        hf_cache_path: Option<PathBuf>,
        matformer_config_path: Option<PathBuf>,
        matformer_slice_name: Option<String>,
    
},
    XLora {

        model_id: Option<String>,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        arch: Option<NormalLoaderType>,
        dtype: ModelDType,
        topology: Option<String>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
        hf_cache_path: Option<PathBuf>,
    
},
    Lora {
        model_id: Option<String>,
        tokenizer_json: Option<String>,
        adapter_model_id: String,
        arch: Option<NormalLoaderType>,
        dtype: ModelDType,
        topology: Option<String>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
        hf_cache_path: Option<PathBuf>,
    },
    GGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    XLoraGGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    LoraGGUF {
        tok_model_id: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        adapters_model_id: String,
        order: String,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    GGML {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        gqa: usize,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    XLoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
        gqa: usize,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    LoraGGML {
        tok_model_id: Option<String>,
        tokenizer_json: Option<String>,
        quantized_model_id: String,
        quantized_filename: String,
        adapters_model_id: String,
        order: String,
        gqa: usize,
        dtype: ModelDType,
        topology: Option<String>,
        max_seq_len: usize,
        max_batch_size: usize,
    },
    VisionPlain {

        model_id: String,
        tokenizer_json: Option<String>,
        arch: Option<VisionLoaderType>,
        dtype: ModelDType,
        topology: Option<String>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        max_edge: Option<u32>,
        calibration_file: Option<PathBuf>,
        imatrix: Option<PathBuf>,
        max_seq_len: usize,
        max_batch_size: usize,
        max_num_images: usize,
        max_image_length: usize,
        hf_cache_path: Option<PathBuf>,
        matformer_config_path: Option<PathBuf>,
        matformer_slice_name: Option<String>,
    
},
    DiffusionPlain {
        model_id: String,
        arch: DiffusionLoaderType,
        dtype: ModelDType,
    },
    Speech {
        model_id: String,
        dac_model_id: Option<String>,
        arch: SpeechLoaderType,
        dtype: ModelDType,
    },
    MultiModel {
        config: String,
        default_model_id: Option<String>,
    },
    Embedding {
        model_id: String,
        tokenizer_json: Option<String>,
        arch: Option<EmbeddingLoaderType>,
        dtype: ModelDType,
        topology: Option<String>,
        write_uqff: Option<PathBuf>,
        from_uqff: Option<String>,
        hf_cache_path: Option<PathBuf>,
    },

}

Variants

Toml

Select the model from a toml file

Fields

file: String

.toml file containing the selector configuration.

Run

Select a model for running via auto loader

Fields

model_id: String

Model ID to load from. May be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If specified, it is used over any remote file.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

organization: Option<IsqOrganization>

ISQ organization: default or moqe.

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;).

imatrix: Option<PathBuf>

.imatrix file to enhance GGUF quantizations with.

calibration_file: Option<PathBuf>

Generate and utilize an imatrix to enhance GGUF quantizations.

max_edge: Option<u32>

Automatically resize and pad images to this maximum edge length. Aspect ratio is preserved. Only supported on specific vision models.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

max_num_images: Option<usize>

Maximum prompt number of images to expect for this model. This affects automatic device mapping but is not a hard limit. Only supported on specific vision models.

max_image_length: Option<usize>

Maximum expected image size will have this edge length on both edges. This affects automatic device mapping but is not a hard limit. Only supported on specific vision models.

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally.

matformer_config_path: Option<PathBuf>

Path to local Matryoshka Transformer configuration CSV file

matformer_slice_name: Option<String>

Name of the Matryoshka Transformer slice to use

Plain

Select a plain model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

arch: Option<NormalLoaderType>

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

organization: Option<IsqOrganization>

ISQ organization: default or moqe (Mixture of Quantized Experts: https://arxiv.org/abs/2310.02410).

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;)

imatrix: Option<PathBuf>

.imatrix file to enhance GGUF quantizations with. Incompatible with --calibration-file/-c

calibration_file: Option<PathBuf>

Generate and utilize an imatrix to enhance GGUF quantizations. Incompatible with --imatrix/-i

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally

matformer_config_path: Option<PathBuf>

Path to local Matryoshka Transformer configuration CSV file

matformer_slice_name: Option<String>

Name of the Matryoshka Transformer slice to use

XLora

Select an X-LoRA architecture

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

arch: Option<NormalLoaderType>

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;).

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally

Lora

Select a LoRA architecture

Fields

model_id: Option<String>

Force a base model ID to load from instead of using the ordering file. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

adapter_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

arch: Option<NormalLoaderType>

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;).

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally

GGUF

Select a GGUF model.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename(s). May be a single filename, or use a delimiter of “ “ (a single space) for multiple files.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

XLoraGGUF

Select a GGUF model with X-LoRA.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename(s). May be a single filename, or use a delimiter of “ “ (a single space) for multiple files.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

LoraGGUF

Select a GGUF model with LoRA.

Fields

tok_model_id: Option<String>

tok_model_id is the local or remote model ID where you can find a tokenizer_config.json file. If the chat_template is specified, then it will be treated as a path and used over remote files, removing all remote accesses.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename(s). May be a single filename, or use a delimiter of “ “ (a single space) for multiple files.

adapters_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

GGML

Select a GGML model.

Fields

tok_model_id: String

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename.

gqa: usize

GQA value

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

XLoraGGML

Select a GGML model with X-LoRA.

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename.

xlora_model_id: String

Model ID to load X-LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached. This makes the maximum running sequences 1.

gqa: usize

GQA value

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

LoraGGML

Select a GGML model with LoRA.

Fields

tok_model_id: Option<String>

Model ID to load the tokenizer from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: String

Quantized model ID to find the quantized_filename. This may be a HF hub repo or a local path.

quantized_filename: String

Quantized filename.

adapters_model_id: String

Model ID to load LoRA from. This may be a HF hub repo or a local path.

order: String

Ordering JSON file

gqa: usize

GQA value

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

VisionPlain

Select a vision plain model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

arch: Option<VisionLoaderType>

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;).

max_edge: Option<u32>

Automatically resize and pad images to this maximum edge length. Aspect ratio is preserved. This is only supported on the Qwen2-VL and Idefics models. Others handle this internally.

calibration_file: Option<PathBuf>

Generate and utilize an imatrix to enhance GGUF quantizations.

imatrix: Option<PathBuf>

.cimatrix file to enhance GGUF quantizations with. This must be a .cimatrix file.

max_seq_len: usize

Maximum prompt sequence length to expect for this model. This affects automatic device mapping but is not a hard limit.

max_batch_size: usize

Maximum prompt batch size to expect for this model. This affects automatic device mapping but is not a hard limit.

max_num_images: usize

Maximum prompt number of images to expect for this model. This affects automatic device mapping but is not a hard limit.

max_image_length: usize

Maximum expected image size will have this edge length on both edges. This affects automatic device mapping but is not a hard limit.

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally

matformer_config_path: Option<PathBuf>

Path to local Matryoshka Transformer configuration CSV file

matformer_slice_name: Option<String>

Name of the Matryoshka Transformer slice to use

DiffusionPlain

Select a diffusion model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

arch: DiffusionLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

Speech

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

dac_model_id: Option<String>

DAC Model ID to load from. If not provided, this is automatically downloaded from the default path for the model. This may be a HF hub repo or a local path.

arch: SpeechLoaderType

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

MultiModel

Select multi-model mode with configuration file

Fields

config: String

Multi-model configuration file path (JSON format)

default_model_id: Option<String>

Default model ID to use when no model is specified in requests

Embedding

Select an embedding model, without quantization or adapters

Fields

model_id: String

Model ID to load from. This may be a HF hub repo or a local path.

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

arch: Option<EmbeddingLoaderType>

The architecture of the model.

dtype: ModelDType

Model data type. Defaults to auto.

topology: Option<String>

Path to a topology YAML file.

write_uqff: Option<PathBuf>

UQFF path to write to.

from_uqff: Option<String>

UQFF path to load from. If provided, this takes precedence over applying ISQ. Specify multiple files using a semicolon delimiter (;)

hf_cache_path: Option<PathBuf>

Cache path for Hugging Face models downloaded locally