Trait mistralrs_core::Loader

pub trait Loader {
    // Required methods
    fn download_model(
        &self,
        revision: Option<String>,
        token_source: TokenSource
    ) -> Result<Box<dyn ModelPaths>>;
    fn _setup_model(
        &self,
        paths: &dyn ModelPaths,
        dtype: Option<DType>,
        device: &Device
    ) -> Result<Box<Mutex<dyn Pipeline + Send + Sync>>>;
    fn get_id(&self) -> &str;
    fn get_kind(&self) -> ModelKind;

    // Provided method
    fn load_model(
        &self,
        revision: Option<String>,
        token_source: TokenSource,
        dtype: Option<DType>,
        device: &Device
    ) -> Result<Box<Mutex<dyn Pipeline + Send + Sync>>> { ... }
}

Encapsulate downloading and setting up the model. The load_model method is used to create the pipeline.

Required Methods

fn download_model( &self, revision: Option<String>, token_source: TokenSource ) -> Result<Box<dyn ModelPaths>>

fn _setup_model( &self, paths: &dyn ModelPaths, dtype: Option<DType>, device: &Device ) -> Result<Box<Mutex<dyn Pipeline + Send + Sync>>>

fn get_id(&self) -> &str

fn get_kind(&self) -> ModelKind

Provided Methods

fn load_model( &self, revision: Option<String>, token_source: TokenSource, dtype: Option<DType>, device: &Device ) -> Result<Box<Mutex<dyn Pipeline + Send + Sync>>>

If revision is None, then it defaults to main. If dtype is None, then it defaults to the model default (usually F32).