Trait mistralrs_core::ModelPaths

pub trait ModelPaths {
    // Required methods
    fn get_weight_filenames(&self) -> &[PathBuf];
    fn get_config_filename(&self) -> &PathBuf;
    fn get_tokenizer_filename(&self) -> &PathBuf;
    fn get_template_filename(&self) -> &PathBuf;
    fn get_adapter_filenames(&self) -> &Option<Vec<(String, PathBuf)>>;
    fn get_adapter_configs(
        &self
    ) -> &Option<Vec<((String, String), LoraConfig)>>;
    fn get_classifier_path(&self) -> &Option<PathBuf>;
    fn get_classifier_config(&self) -> &Option<XLoraConfig>;
    fn get_ordering(&self) -> &Option<Ordering>;
    fn get_gen_conf_filename(&self) -> Option<&PathBuf>;
    fn get_lora_preload_adapter_info(
        &self
    ) -> &Option<HashMap<String, (PathBuf, LoraConfig)>>;
}

ModelPaths abstracts the mechanism to get all necessary files for running a model. For example LocalModelPaths implements ModelPaths when all files are in the local file system.

Required Methods

fn get_weight_filenames(&self) -> &[PathBuf]

Model weights files (multiple files supported).

fn get_config_filename(&self) -> &PathBuf

Retrieve the PretrainedConfig file. See: https://huggingface.co/docs/transformers/v4.40.2/en/main_classes/configuration#transformers.PretrainedConfig

fn get_tokenizer_filename(&self) -> &PathBuf

A serialised tokenizers.Tokenizer HuggingFace object. See: https://huggingface.co/docs/transformers/v4.40.2/en/main_classes/tokenizer

fn get_template_filename(&self) -> &PathBuf

Content expected to deserialize to [ChatTemplate].

fn get_adapter_filenames(&self) -> &Option<Vec<(String, PathBuf)>>

Optional adapter files. (String, PathBuf) is of the form (id name, path).

fn get_adapter_configs(&self) -> &Option<Vec<((String, String), LoraConfig)>>

Configuration of optional adapters. (String, String) is of the form (id name, name).

fn get_classifier_path(&self) -> &Option<PathBuf>

Filepath for the XLORA classifier

fn get_classifier_config(&self) -> &Option<XLoraConfig>

XLoraConfig for the XLORA classifier

fn get_ordering(&self) -> &Option<Ordering>

Return the defined ordering of adapters and layers within the model.

fn get_gen_conf_filename(&self) -> Option<&PathBuf>

Filepath for general model configuration.

fn get_lora_preload_adapter_info( &self ) -> &Option<HashMap<String, (PathBuf, LoraConfig)>>