mistralrs_core

Trait ModelPaths

pub trait ModelPaths:
    AsAny
    + Debug
    + Send
    + Sync {
    // Required methods
    fn get_weight_filenames(&self) -> &[PathBuf];
    fn get_config_filename(&self) -> &PathBuf;
    fn get_tokenizer_filename(&self) -> &PathBuf;
    fn get_template_filename(&self) -> &Option<PathBuf>;
    fn get_gen_conf_filename(&self) -> Option<&PathBuf>;
    fn get_preprocessor_config(&self) -> &Option<PathBuf>;
    fn get_processor_config(&self) -> &Option<PathBuf>;
    fn get_chat_template_explicit(&self) -> &Option<PathBuf>;
    fn get_adapter_paths(&self) -> &AdapterPaths;
    fn get_modules(&self) -> Option<&[EmbeddingModulePaths]>;
}

ModelPaths abstracts the mechanism to get all necessary files for running a model. For example LocalModelPaths implements ModelPaths when all files are in the local file system.

Required Methods

fn get_weight_filenames(&self) -> &[PathBuf]

Model weights files (multiple files supported).

fn get_config_filename(&self) -> &PathBuf

Retrieve the PretrainedConfig file.

fn get_tokenizer_filename(&self) -> &PathBuf

A serialised tokenizers.Tokenizer HuggingFace object.

fn get_template_filename(&self) -> &Option<PathBuf>

File where the content is expected to deserialize to ChatTemplate.

fn get_gen_conf_filename(&self) -> Option<&PathBuf>

Filepath for general model configuration.

fn get_preprocessor_config(&self) -> &Option<PathBuf>

Get the preprocessor config (for the vision models). This is used to pre process images.

fn get_processor_config(&self) -> &Option<PathBuf>

Get the processor config (for the vision models). This is primarily used for the chat template.

fn get_chat_template_explicit(&self) -> &Option<PathBuf>

Get the explicit chat template.

fn get_adapter_paths(&self) -> &AdapterPaths

Get adapter paths.

fn get_modules(&self) -> Option<&[EmbeddingModulePaths]>

Get embedding model modules.json compatible with sentence-transformers